Refactor webpack code in server.js into dev.js; change package.json server route
