+++ package.json
@@ -9,8 +9,7 @@
   },
   "scripts": {
     "test": "jest",
-    "build": "webpack --config ./webpack.config.js",
-    "dev": "npm build && nodemon server.js --ignore /client",
+    "dev": "nodemon server/server.js --ignore /client",
     "start": "node server/server.js",
     "heroku-postbuild": "webpack -p --config ./webpack.prod.config.js --progress",
     "lint": "eslint client server"
