+++ server/dev.js
@@ -1,0 +1,18 @@
+module.exports.webpack = app => {
+  const webpack = require('webpack');
+  const webpackDevMiddleware = require('webpack-dev-middleware');
+  const webpackHotMiddleware = require('webpack-hot-middleware');
+  const config = require('../webpack.config.js');
+  const compiler = webpack(config);
+
+  app.use(webpackHotMiddleware(compiler));
+  app.use(webpackDevMiddleware(compiler, {
+    publicPath: '/',
+    stats: {
+      colors: true,
+    },
+    hot: true,
+    noInfo: true,
+  }));
+
+}
